issue #7 (refactor - api): renamed a lot of files
