+++ src/dao/impl/postgres/postgres_chapters.dao.impl.ts
@@ -1,6 +1,6 @@
 import { Knex } from "knex";
 import db from "../../../database/database";
-import { IChapterDAO } from "../../chapter_dao";
+import { IChapterDAO } from "../../chapter.dao.interface";
 import { IChapter } from "../../../models/chapter";
 
 export class PostgresChapterDAO implements IChapterDAO {
